data_generation/requirements.txt
mimesis
psycopg2-binary==2.9.9